requirements.txt
fastapi
uvicorn
openai
faiss-cpu
tqdm
requests
